requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
fastapi
uvicorn

# AI and ML libraries
openai>=1.0.0
google-cloud-aiplatform
google-generativeai
vertexai

# Google Cloud services
google-cloud-storage

# Vector database and embeddings
chromadb
langchain
langchain-community
llama-index

# Data processing
pandas
numpy

# Document generation
weasyprint
python-docx

# Utilities
requests
user-agent
python-multipart
aiofiles
pydantic>=2.0.0

# For agent functionality
smolagents

# Additional for cloud deployment
google-auth
google-auth-oauthlib
google-auth-httplib2